adapted to Approach 3
currently working locally because the server is down (have commented out somethings and will get back to them later on).
have adapted the database, the endpoints and the UI to Approach 3.
have added 2 new pages which I will get back to as soon as possible, adapting their dummy endpoints, etc (add and preview images).
have adapted the ERD as well, to have a visual representation of the system's database.